revert previous commit
git-svn-id: https://svn.salilab.org/imp/trunk@1379 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ kernel/include/base_types.h
@@ -20,7 +20,7 @@ IMP_BEGIN_NAMESPACE
  */
 
 //! Basic floating-point value (could be float, double...)
-typedef double Float;
+typedef float Float;
 
 //! Basic integer value
 typedef int Int;
